+++ sovtimer/views.py
@@ -1,5 +1,5 @@
 """
-the views
+The views
 """
 
 # Standard Library
@@ -7,7 +7,8 @@ import datetime as dt
 
 # Django
 from django.contrib.auth.decorators import login_required, permission_required
-from django.http import JsonResponse
+from django.core.handlers.wsgi import WSGIRequest
+from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render
 from django.utils.translation import gettext_lazy as _
 
@@ -27,13 +28,9 @@ from sovtimer.utils import LoggerAddTag
 logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
-# module constants
-CAMPAIGN_TREND_CACHE_TIME = 30
-
-
 @login_required
 @permission_required("sovtimer.basic_access")
-def dashboard(request):
+def dashboard(request: WSGIRequest) -> HttpResponse:
     """
     Index view
     """
@@ -45,7 +42,7 @@ def dashboard(request):
     return render(request, "sovtimer/dashboard.html", context)
 
 
-def dashboard_data(request) -> JsonResponse:
+def dashboard_data(request: WSGIRequest) -> JsonResponse:
     """
     ajax call
     get dashboard data
@@ -53,7 +50,7 @@ def dashboard_data(request) -> JsonResponse:
     :return:
     """
 
-    data = list()
+    data = []
 
     sovereignty_campaigns = Campaign.objects.select_related(
         "structure",
@@ -81,41 +78,28 @@ def dashboard_data(request) -> JsonResponse:
             )
 
             defender_name_html = (
-                '<a href="{defender_url}" '
-                'target="_blank" rel="noopener noreferer">'
-                '<img class="aa-sovtimer-entity-logo-left" '
-                'src="{defender_logo_url}" alt="{defender_name}">'
-                "{defender_name}</a>".format(
-                    defender_logo_url=defender_logo_url,
-                    defender_url=defender_url,
-                    defender_name=defender_name,
-                )
+                f'<a href="{defender_url}" target="_blank" rel="noopener noreferer">'
+                f'<img class="aa-sovtimer-entity-logo-left" src="{defender_logo_url}" '
+                f'alt="{defender_name}">{defender_name}</a>'
             )
 
             # solar system
             solar_system_name = campaign.structure.solar_system.name
-            solar_system_url = (
-                dotlan_region_url(
-                    campaign.structure.solar_system.eve_constellation.eve_region
-                )
-                + "/"
-                + campaign.structure.solar_system.name
+            region_url = dotlan_region_url(
+                campaign.structure.solar_system.eve_constellation.eve_region
             )
+            campaign_system_name = campaign.structure.solar_system.name
+            solar_system_url = f"{region_url}/{campaign_system_name}"
             solar_system_name_html = (
-                '<a href="{solar_system_url}" '
-                'target="_blank" rel="noopener noreferer">'
-                "{solar_system_name}</a>".format(
-                    solar_system_url=solar_system_url,
-                    solar_system_name=solar_system_name,
-                )
+                f'<a href="{solar_system_url}" target="_blank" '
+                f'rel="noopener noreferer">{campaign_system_name}</a>'
             )
 
             # constellation
             constellation_name = campaign.structure.solar_system.eve_constellation.name
             constellation_name_html = (
-                '<a href="//evemaps.dotlan.net/search?q={constellation_name}" '
-                'target="_blank" rel="noopener noreferer">'
-                "{constellation_name}</a>".format(constellation_name=constellation_name)
+                f'<a href="//evemaps.dotlan.net/search?q={constellation_name}" '
+                f'target="_blank" rel="noopener noreferer">{constellation_name}</a>'
             )
 
             # region
@@ -126,11 +110,8 @@ def dashboard_data(request) -> JsonResponse:
                 campaign.structure.solar_system.eve_constellation.eve_region
             )
             region_name_html = (
-                '<a href="{region_url}" '
-                'target="_blank" rel="noopener noreferer">{region_name}</a>'.format(
-                    region_url=region_url,
-                    region_name=region_name,
-                )
+                f'<a href="{region_url}" target="_blank" '
+                f'rel="noopener noreferer">{region_name}</a>'
             )
 
             # activity defense multiplier
@@ -152,9 +133,10 @@ def dashboard_data(request) -> JsonResponse:
             # campaign progress field
             campaign_progress_previous = campaign.progress_previous
             campaign_progress_previous_percentage = campaign_progress_previous * 100
-            campaign_progress_previous_percentage_formatted = "{:.0f}%".format(
-                campaign_progress_previous_percentage
+            campaign_progress_previous_percentage_formatted = (
+                f"{campaign_progress_previous_percentage:.0f}%"
             )
+
             if campaign_progress_previous_percentage < 10:
                 campaign_progress_previous_percentage_formatted = (
                     "0" + campaign_progress_previous_percentage_formatted
@@ -162,8 +144,8 @@ def dashboard_data(request) -> JsonResponse:
 
             campaign_pogress_current = campaign.progress_current
             campaign_pogress_current_percentage = campaign_pogress_current * 100
-            campaign_pogress_current_percentage_formatted = "{:.0f}%".format(
-                campaign_pogress_current_percentage
+            campaign_pogress_current_percentage_formatted = (
+                f"{campaign_pogress_current_percentage:.0f}%"
             )
             if campaign_pogress_current_percentage < 10:
                 campaign_pogress_current_percentage_formatted = (
@@ -176,40 +158,35 @@ def dashboard_data(request) -> JsonResponse:
             if remaining_time_in_seconds < 0:
                 active_campaign = _("Yes")
 
+                progress_text = _("Neither side has made any progress yet")
                 campaign_progress_icon = (
-                    '<i class="material-icons aa-sovtimer-trend '
-                    'aa-sovtimer-trend-flat" '
-                    'title="{title_text}">trending_flat</i>'.format(
-                        title_text=_("Neither side has made any progress yet")
-                    )
+                    f"<i "
+                    f'class="material-icons aa-sovtimer-trend aa-sovtimer-trend-flat" '
+                    f'title="{progress_text}">trending_flat</i>'
                 )
 
                 if campaign_progress_previous < campaign_pogress_current:
+                    defender_progress_text = _("Defenders making progress")
                     campaign_progress_icon = (
-                        '<i class="material-icons aa-sovtimer-trend '
-                        'aa-sovtimer-trend-up" '
-                        'title="{title_text}">trending_up</i>'.format(
-                            title_text=_("Defenders making progress")
-                        )
+                        f"<i "
+                        f'class="material-icons aa-sovtimer-trend aa-sovtimer-trend-up" '
+                        f'title="{defender_progress_text}">trending_up</i>'
                     )
 
                 if campaign_progress_previous > campaign_pogress_current:
+                    attacker_progress_text = _("Attackers making progress")
                     campaign_progress_icon = (
-                        '<i class="material-icons aa-sovtimer-trend '
-                        'aa-sovtimer-trend-down" '
-                        'title="{title_text}">trending_down</i>'.format(
-                            title_text=_("Attackers making progress")
-                        )
+                        f"<i "
+                        f'class="material-icons aa-sovtimer-trend aa-sovtimer-trend-down" '
+                        f'title="{attacker_progress_text}">trending_down</i>'
                     )
 
+                constellation_id = campaign.structure.solar_system.eve_constellation.id
+                zkb_href = f"https://zkillboard.com/constellation/{constellation_id}/"
+                zkb_icon = '<img src="/static/sovtimer/images/zkillboard.png">'
                 constellation_killboard_link = (
-                    '<a href="{zkb_url}{constellation_id}/" '
-                    'target="_blank" rel="noopener noreferer" '
-                    'class="aa-sov-timer-zkb-icon">{zkb_icon}</a>'.format(
-                        zkb_url="https://zkillboard.com/constellation/",
-                        constellation_id=campaign.structure.solar_system.eve_constellation.id,
-                        zkb_icon='<img src="/static/sovtimer/images/zkillboard.png">',
-                    )
+                    f'<a href="{zkb_href}" target="_blank" rel="noopener noreferer" '
+                    f'class="aa-sov-timer-zkb-icon">{zkb_icon}</a>'
                 )
 
                 campaign_progress_html = (